# backend/internal/database/model/resource.go
package model

import "time"

type Resource struct {
	ID        int64     `json:"id" gorm:"primarykey;autoIncrement"`
	IsLocked  bool      `json:"is_locked" gorm:"is_locked"`
	Kind      string    `json:"kind" gorm:"kind"`
	Value     string    `json:"value" gorm:"value"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

func (Resource) TableName() string {
	return "authz_resources"
}